+++ packages/fxa-shared/payments/iap/google-play/user-manager.ts
@@ -4,6 +4,7 @@
 
 import { CollectionReference } from '@google-cloud/firestore';
 import { ILogger } from 'fxa-shared/log';
+import { PurchaseManager } from './purchase-manager';
 import {
   GOOGLE_PLAY_FORM_OF_PAYMENT,
   SubscriptionPurchase,
@@ -12,35 +13,19 @@ import { SkuType } from './types/purchases';
 
 export class UserManager {
   constructor(
-    public purchasesDbRef: CollectionReference,
+    protected purchasesDbRef: CollectionReference,
+    protected purchaseManager: PurchaseManager,
     protected log: ILogger
   ) {}
 
-  async querySubscriptions(
+  async queryCurrentSubscriptions(
     userId: string,
     sku?: string,
     packageName?: string
   ): Promise<Array<SubscriptionPurchase>> {
-    // Create query to fetch possibly active subscriptions from Firestore
-    let query = this.buildSubscriptionQuery(userId, sku, packageName);
-
-    // Do fetch possibly active subscription from Firestore
-    const result = await query.get();
-
     const purchaseList = new Array<SubscriptionPurchase>();
-    for (const purchaseRecordSnapshot of result.docs) {
-      purchaseList.push(
-        SubscriptionPurchase.fromFirestoreObject(purchaseRecordSnapshot.data())
-      );
-    }
-    return purchaseList;
-  }
 
-  protected buildSubscriptionQuery(
-    userId: string,
-    sku: string | undefined,
-    packageName: string | undefined
-  ) {
+    // Create query to fetch possibly active subscriptions from Firestore
     // Create query to fetch possibly active subscriptions from Firestore
     let query = this.purchasesDbRef
       .where('formOfPayment', '==', GOOGLE_PLAY_FORM_OF_PAYMENT)
@@ -55,6 +40,44 @@ export class UserManager {
     if (packageName) {
       query = query.where('packageName', '==', packageName);
     }
-    return query;
+
+    // Do fetch possibly active subscription from Firestore
+    const queryResult = await query.get();
+
+    // Loop through these subscriptions and filter those that are indeed active
+    for (const purchaseRecordSnapshot of queryResult.docs) {
+      let purchase: SubscriptionPurchase =
+        SubscriptionPurchase.fromFirestoreObject(purchaseRecordSnapshot.data());
+
+      if (
+        !purchase.isEntitlementActive() &&
+        !purchase.isAccountHold() &&
+        !purchase.isPaused()
+      ) {
+        // If a subscription purchase record in Firestore indicates says that it has expired,
+        // and we haven't confirmed that it's in Account Hold,
+        // and we know that its status could have been changed since we last fetch its details,
+        // then we should query Play Developer API to get its latest status
+        this.log.info('queryCurrentSubscriptions.cache.update', {
+          purchaseToken: purchase.purchaseToken,
+        });
+        purchase = await this.purchaseManager.querySubscriptionPurchase(
+          purchase.packageName,
+          purchase.sku,
+          purchase.purchaseToken
+        );
+      }
+
+      // Add the updated purchase to list to returned to clients
+      if (
+        purchase.isEntitlementActive() ||
+        purchase.isAccountHold() ||
+        purchase.isPaused()
+      ) {
+        purchaseList.push(purchase);
+      }
+    }
+
+    return purchaseList;
   }
 }
